requirements.txt
openai-whisper
ffmpeg-python
torch
pydub
noisereduce
